lsp-devtools: move cli command definitions to cli module
In an attempt to draw a stronger line between "cli code"  and "library code", this commit
moves all the argparse code for each command into the `.cli` module.
